refactor: make rvr use air index enum end-to-end
